Bug fix: when undoing the creation of an index, a NPE would occur
It appears that we will run into trouble in our (graphical) edit parts when
handling upon each individual property change event; in doing so, we will
be exposed to temporary inconsistencies in the schema model (like an area
specification that is not yet or no longer associated with an area).

The outline related (tree) edit parts follow command stack changes but
this is too coarse.

It would be better to respect the fact that a command corresponds to a
unit of work and only act when a unit of work is executed, undone or
redone...