Add model change annotations (model change notification redesign).
Model changes are to be controlled in a better way; components
interested in model changes (edit parts, property sections, ...) should
be informed about model changes in a controlled way, i.e. the model 
should always be in a consistent state when informing a component of a 
model change; this is difficult, if not impossible to achieve with EMF
model change events.

Each command executed on the command stack and for which model changes
are to be dispatched to interested components, is assigned a 'model 
change category' and depending on the model change category, 1 or more
fields of the command instance have to be annotated as well (TODO: 
document all annotations so that it is clear which annotation is 
applicable for which model change category).