Hookup the model change notification mechanism in the SchemaEditor (model change notification redesign).
Model changes are to be controlled in a better way; components
interested in model changes (edit parts, property sections, ...) should
be informed about model changes in a controlled way, i.e. the model 
should always be in a consistent state when informing a component of a 
model change; this is difficult, if not impossible to achieve with EMF
model change events.

A SchemaEditor creates its own ModelChangeDispatcher; it still listens
for command stack events, and will dispatch each event to the
ModelChangeDispatcher from now on.  The SchemaEditor should be optimized
to use this mechanism itself in a couple of places (e.g. when the zoom
level changes).