Split the following queries into several pieces joined by the UNION keyword, depending on the number of dbkeys provided :
- BaseRecordSynonymListQuery
- ElementListQuery
- ElementCommentListQuery

Each query part now contains a configurable maximum number of dbkeys
(defaulting to 47, for that number seems to be the point after which
query performance become awful); the query is no longer composed of a
single long dbkey list in the WHERE clause.  Whereas the query result is
the same, performance seems to improve a lot when working this way.

Note that we needed to add an ORDER BY clause for the ElementListQuery
because rows were not always returned in the order of the set occurrence
for set RCDSYN-NAMESYN (more specifically: sometimes the reverse order
is returned); we did the same for the ElementCommentListQuery although
we did not run into trouble with that query.  It seems that it is good
practice to always (try to) provide an ORDER BY clause (i.e. if the
query result order matters).