1. Replace the appliesTo(EObject) method in the ModelChangeContext class by individual methods, also called 'appliesTo', but each accepting a specific type as its only parameter:
- ConnectionPart
- Connector
- Guide
- MemberRole
- Ruler
- SchemaArea
- SchemaRecord
- Set

Note that some types can no longer be passed as an argument to the
'appliesTo' method(s): ConnectionLabel, DiagramData, DiagramLabel,
Schema and SystemOwner.

2. Replace the putContextData(EObject) method in the ModelChangeContext
class by individual methods, also called 'putContextData', but each
accepting a specific type as its only parameter:

- ConnectionPart
- Connector
- DiagramNode (a ConnectionLabel, Connector, DiagramLabel, SchemaRecord
or SystemOwner)
- Guide
- MemberRole
- Ruler
- SchemaArea
- SchemaRecord
- Set

Note that 2 types can no longer be passed as an argument to the
'putContextData' method(s): DiagramData and Schema.

Finally, a putContextData method has been added, specifically (and only)
for SET_PROPERTY model change types, accepting 2 parameters: a model
object and the feature; the supported model types are:

- DiagramData
- DiagramLabel
- Schema
- SchemaArea
- SchemaRecord
- Set

The above list corresponds to the types of defined attributes based
properties sections (Properties view).

3. The CALC key duplicates option is no longer directly selectable in
the Properties view (record CALC section) but is replaced by a
hyperlink, that links to the record location mode dialog.

4. The symbolic displacement and displacement page count are no longer
directly editable in the Properties view (record VIA section) but are
replaced by a hyperlink, that links to the record location mode dialog.

5. Fix issue with symbolic displacement and displacement page count;
these are mutually exclusive but could be set both (record location mode
dialog).

6. Enhance the hyperlink in the Properties view so that empty values
appear as '[...]'.