Add TODO items regarding the delete index command.
In its current implementation, the delete index command also removes an
obsolete area and/or procedures.  Since we now have dedicated commands
to remove an area or procedure, it would be better to have a compound
command taking care of these aspects rather than duplicating that
functionality in the delete index command.  We will definitely have to
tackle this issue when we allow records to be deleted when they
participate in sets (a record can currently only be removed if it does
NOT participate in any set, either as an owner or member).