Fix issue with moving a group of diagram nodes (records, indexes, ...).
It appears that the DragEditPartsTracker class creates a compound
command when more than 1 diagram nodes are being moved; we do NOT have
control over this in one of our edit policies.

As a solution, we create a last-minute model change context when
dispatching the model change event a compound command under that
circumstance.