Make the Update... menu item available in the Project Explorer.
Until now, this menu item was only available in the Package Explorer.

When not available (e.g. when right-clicking on a folder in a project,
the menu item will no longer be visible; it was disabled, but from now
on it will be invisible).

While we were there we fixed a wrong property test for this
functionality; the property 'SchemaFileSelected' should be tested and
not 'recordSelected' - this bug was probably introduced while creating
the Import Record Elements... menu item (Diagram menu).