Bug fix: import from dictionary fails with internal error.
We would go look for an SDR-042 in the wrong record synonym.  The
following error appears in the log :

Caused by: java.lang.RuntimeException: internal error: no NAMESYN-083
for an SDR-042 with dbkey X'06594502' under RCDSYN-079=X'06594502'

Mind that both dbkeys mentioned are the same; this is another problem
(which we fixed as well).

at Rcdsyn_079.getNamesyn_083(Rcdsyn_079.java:57)