Bug fix:  when changing an area name via the area dialog, not all record figures show the new area name.
This bug was introduced while creating the revised model change
notification system.

It is possible that, together with changing an area specification, the
area is being renamed as well; the record whose area specification is
changed will be refreshed already, but we need to make sure that the new
area name replaces the old one for all other records contained in that
area (we might unnecessarily do a refresh, but that's better than
missing an area rename).