Bug fix: set order dialog cannot be opened for indexed sets ordered by dbkey.
Opening the set order dialog for indexed sets ordered by dbkey results
in a NPE because it is not taken into account that key elements can
represent a dbkey; an actual element is always expected.  Moreover, the
commands responsible for modifying the model in the case sort keys are
involved experience a similar problem :  an actual element is always
expected there.  This is now fixed.