Bug fix: wrong record syntax when no 'SHARE STRUCTURE' present during import.
When importing a diagram from schema syntax and no 'SHARE STRUCTURE'
clause is present for a record, the syntax tab in the Properties view
will show 'SHARE STRUCTURE OF RECORD null VERSION 0 ' and the record
information will be incorrect in the Dictionaries tab.  From now on,
that missing information will default to the record name and the
version(s) will be set to 1.