Change model for VSAM support.
Added the following location modes :  VSAM and VSAM_CALC.
Added the UNORDERED duplicates option.
Added the VsamType class, referenced by the SchemaRecord class.
Added the VsamLengthType enum.
Added the VSAM_INDEX set mode; in order to visualize this in a diagram,
the VsamIndex class (which subclasses DiagramNode) is provided.

As an extra: added a boolean indicating whether a record is VSAM, as
well as a boolean indicating whether a set has mode VSAM_INDEX (plus,
for sets and for completeness we added a boolean indicating whether the
set is chained and one for the indexed set mode; this is more in
accordance to the ones available in records).