Bug fix: renaming an area throws a NPE & destroys model change event processing
This bug was introduced with commit 4114a8e (Bug fix:  when changing an
area name via the area dialog, not all record figures show the new area
name.).  What we forgot while solving that problem, is that areas can
have a mix of records AND indexes; moreover, that commit didn't solve
the rename problem for indexes).

The fact that a NPE can prevent further model change notifications to be
dealt with normally has to taken care of by surrounding the invocation
of the before- and after model change methods on the listeners with a
try/catch block.