Bug fix: import from syntax does NOT detect default WITHIN AREA clause (indexes)
For system owned indexed sets, the WITHIN AREA clause is optional and
when omitted, the SR7 record is stored in the same area as the member
record.  The punched syntax WILL contain a WITHIN AREA clause referring
to the member record's area BUT this clause is commented out, which
prevents the syntax import tool to detect it.