Bug fix: creating a user owned indexed set does NOT create indexed set mode spec
The result of that situation is that generating the schema syntax fails
with a NPE because for indexed sets, whether user or system owned, an
indexed set mode specification should ALWAYS be present.

Along the way we extend the test case for creating a system owned
indexed set because there is no validation of the indexed set mode
specification (which is created correctly btw).