Eliminate the 'redefines check (not for CALC keys)' on schema import
The CA IDMS documentation specifies the following restriction on
elements that can be used in a set's sort key (i.e. elements conforming
to the condition below are not suitable for inclusion in a sort key):

- No element that redefines another element or is subordinate to an
element that redefines another element can be used in the sort control
element.

Apparently, schemas exist containing set sort keys that contain such
elements, so this constraint was removed.