Go for a POM-less Tycho build
Switch to Tycho 0.24.0 and remove all unnecessary pom.xml files; the
only ones remaining are the one in the parent project and in the update
site project (i.e. the pom.xml files for bundles and features have
become unnecessary).

This requires Maven 3.3 or higher.  See
https://wiki.eclipse.org/Tycho/Release_Notes/0.24.

By doing this, incrementing versions becomes a lot simpler.