Stop virtually locking connection part endpoints
For connection parts without a source and/or target endpoint, a kind of
'virtual endpoint' is set when creating the owner and member records and
this endpoint location (an offset to the record figure) never changes.
By doing so, it is possible to layout a connection part and get a
different layout when closing and reopening the diagram.