Add support for the .schemadsl file extension
Files can now be saved as .schemadsl; the file then contains DSL instead
of XMI.  The diagram editor is automatically associated to this file
extension.

Open editors for this new file extension are ALWAYS closed upon
workbench shutdown (if dirty the user will be prompted to save or
discard the changes).  The reason for this is that it can take a long
time for files of this type to be opened in the diagram editor (probably
because the file contents needs to be compiled by the Groovy compiler)
and we want to avoid that the workbench restart takes too long.