Bug fix: rulers are ignored in the DSL and are lost when saving to .schemadsl
It appears that the schema DSL syntax builder totally ignores horizontal
and vertical guides and the schema model builder has no support for
guides at all.

Mind that the new DSL syntax (horizontalGuides and verticalGuides, both
under the schema's diagramData) can be confusing because:
- the horizontal guiders are owned by the vertical ruler
- the vertical guides are owned by the horizontal ruler

The DSL syntax expresses the nature of both kind of guides in a good
way, the data model underneath is of less importance here.