Quick fix for procedure call specification issue in .schemadsl files
In Groovy 4, the closure meta class does no longer seem to allow us to
override the call method (which we translate to a call to the
callProcedure method).

This tweak manipulates the .schemadsl file contents when loading it into
the editor: it replaces all occurrences of "call '" by "callProcedure
'". Saving a .schemadsl file will however still result in the original
area/record procedure call specification (i.e. with call and not
callProcedure).